rework getStatistics and getHealthChecks
export cf health fun

all funs linting

fix tests

fix tests
